feat: 스크롤 감지 로직 훅으로 분리 (#79)
